requirements.txt
# LiveKit
livekit>=0.13.0

# Async HTTP/WebSocket
httpx>=0.27.0
websockets>=12.0

# AI Providers
openai>=1.40.0

# VAD
silero-vad>=5.1
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Persistence
redis>=5.0.0
asyncpg>=0.29.0

# Resilience
tenacity>=8.2.0

# Observability
structlog>=24.1.0

# Environment
python-dotenv>=1.0.0
